Used static_cast to reduce the warnings.

+++ src/scripts/InstanceScripts/Instance_Arcatraz.cpp
@@ -183,7 +183,7 @@ class ZerekethAI : public MoonScriptBossAI
             if (!TargetTable.size())
                 return;
 
-            auto random_index = RandomUInt(0, TargetTable.size() - 1);
+            auto random_index = RandomUInt(0, uint32(TargetTable.size() - 1));
             auto random_target = TargetTable[random_index];
 
             if (random_target == nullptr)
