+++ src/scripts/InstanceScripts/Instance_TheShatteredHalls.cpp
@@ -1852,7 +1852,7 @@ class GrandWarlockNethekurseAI : public CreatureAIScript
                 if (!TargetTable.size())
                     return;
 
-                auto random_index = RandomUInt(0, TargetTable.size() - 1);
+                auto random_index = RandomUInt(0, uint32(TargetTable.size() - 1));
                 auto random_target = TargetTable[random_index];
 
                 if (random_target == nullptr)
